#  clusterizer: Iteratively cut KD branches to speed up traversal