--- a/src/clusterizer.cpp
+++ b/src/clusterizer.cpp
@@ -681,15 +681,22 @@ static void kdtreeNearest(KDNode* nodes, unsigned int root, const float* points,
 {
 	const KDNode& node = nodes[root];
 
+	if (node.children == 0)
+		return;
+
 	if (node.axis == 3)
 	{
 		// leaf
+		bool inactive = true;
+
 		for (unsigned int i = 0; i < node.children; ++i)
 		{
 			unsigned int index = nodes[root + i].index;
 
 			if (emitted_flags[index])
 				continue;
+
+			inactive = false;
 
 			const float* point = points + index * stride;
 
@@ -702,6 +709,10 @@ static void kdtreeNearest(KDNode* nodes, unsigned int root, const float* points,
 				limit = distance;
 			}
 		}
+
+		// deactivate leaves that no longer have items to emit
+		if (inactive)
+			nodes[root].children = 0;
 	}
 	else
 	{
@@ -709,6 +720,11 @@ static void kdtreeNearest(KDNode* nodes, unsigned int root, const float* points,
 		float delta = position[node.axis] - node.split;
 		unsigned int first = (delta <= 0) ? 0 : node.children;
 		unsigned int second = first ^ node.children;
+
+		// deactivate branches that no longer have items to emit to accelerate traversal
+		// note that we do this *before* recursing which delays deactivation but keeps tail calls
+		if ((nodes[root + 1 + first].children | nodes[root + 1 + second].children) == 0)
+			nodes[root].children = 0;
 
 		kdtreeNearest(nodes, root + 1 + first, points, stride, emitted_flags, position, result, limit);