On meshes with many topologically disjoint clusters, kdtreeNearest
spends an increasing amount of time going over already emitted
triangles. To fix this, we can detect leaves that have all triangles
already emitted and flag them as inactive using children=0. This, then,
can be propagated to branch nodes using children=0 as a flag as well.

For branches, we do this *before* visiting either node. This makes
deactivation lazier than it strictly speaking needs to be, but it keeps
the existing recursive traversal structure where the second call is a
tail call, which on some models results in extra speedup.

Overall, while on some meshes this is neutral because the KD tree search
is used rarely, on meshes with more seams this can improve the meshlet
building time anywhere from 15% to 40%:

Mesh | Before | After | Speedup
--------|-----------|---------|------------
buddha.obj | 415 ms | 385 ms | 7%
bistro_exterior.obj | 1080 ms | 780 ms | 28%
bistro_interior.obj | 310 ms | 260 ms | 16%
powerplant.obj | 7400 ms | 4170 ms | 43%
rungholt.obj | 3800 ms | 2630 ms | 31%
sanMiguel.obj | 4200 ms | 2600 ms | 38%
thai_buddha.obj | 2750 ms | 2660 ms | 3%
Bistro GLB + lods | 1080 ms | 940 ms | 13%

*This contribution is sponsored by Valve.*